Regression Analysis
The following script uses regression analysis to show the probability of an item being recovered if it was an automobile. Though all property types were initially analyzed, it became clear that returned probability differed by item type, and one could estimate an item's return value if it was an automobile.