Unsupervized Machine Learning
This script uses unsupervised machine learning to break the data into different clustered sets. The analysis broke the data into different clusters based on whether it was returned and how high the initial property value was. Since property type plays a strong role in this and was not captured by the analysis, machine learning was not beneficial to the final project.